pyproject.toml
[project]
name = "multitasking-text-utility"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
]
